구조 조정
1. 세션 map의 lock을 해제하여 배열 형태로 변경.

2. 테스트 이후 성능을 위해 로그 찍는 부분을 비활성화.

+++ 20241113 IOCP/Content.cpp
@@ -6,7 +6,7 @@
 
 UINT32 CPlayer::g_ID = 0;
 
-void ServerContent::OnRecv(int sessionID, CPacket* pPacket)
+void ServerContent::OnRecv(UINT64 sessionID, CPacket* pPacket)
 {
 	// ¿©±â¼­ ÄÁÅÙÃ÷ Ã³¸®°¡ µé¾î°¨.
 	// ~ ÄÁÅÙÃ÷ ~
@@ -15,7 +15,7 @@ void ServerContent::OnRecv(int sessionID, CPacket* pPacket)
 	pIServer->SendPacket(sessionID, pPacket);
 }
 
-void ServerContent::OnAccept(int sessionID)
+void ServerContent::OnAccept(UINT64 sessionID)
 {
 	CPlayer* pPlayer = new CPlayer;
 	pPlayer->sessionID = sessionID;
