+++ 20241113 IOCP/Content.h
@@ -30,10 +30,10 @@ public:
 
 public:
     // 클라이언트로부터 데이터를 수신했을 때 호출되는 함수. 클라이언트에서 정의
-    virtual void OnRecv(int sessionID, CPacket* pPacket) override;
+    virtual void OnRecv(UINT64 sessionID, CPacket* pPacket) override;
 
     // 서버가 세션을 받고, 세션과 연결된 객체를 만들기 위해 컨텐츠에서 호출해야하는 함수
-    virtual void OnAccept(int sessionID) override;
+    virtual void OnAccept(UINT64 sessionID) override;
 
 private:
     std::unordered_map<UINT32, CPlayer*> playerMap;
